refactor: extract rest of the types into separate files and extract enum
