+++ Applications/AdminUi/packages/admin_api_types/lib/src/admin_api_types_base.dart
@@ -1,4 +1,5 @@
 export 'clients/clients.dart';
+export 'enums/enums.dart';
 export 'identities/identities.dart';
 export 'messages/messages.dart';
 export 'metrics/metrics.dart';
