+++ Applications/AdminUi/packages/admin_api_types/lib/src/enums/deletion_process_status.dart
@@ -1,0 +1,18 @@
+import 'package:json_annotation/json_annotation.dart';
+
+enum DeletionProcessStatus {
+  @JsonValue('WaitingForApproval')
+  waitingForApproval,
+
+  @JsonValue('Approved')
+  approved,
+
+  @JsonValue('Cancelled')
+  cancelled,
+
+  @JsonValue('Rejected')
+  rejected,
+
+  @JsonValue('Deleting')
+  deleting,
+}
