+++ Applications/AdminUi/packages/admin_api_types/lib/src/identities/identities.dart
@@ -1,5 +1,6 @@
 export 'identity_deletion_process.dart';
 export 'identity_deletion_process_audit_log_entry.dart';
+export 'identity_deletion_process_detail.dart';
 export 'identity_device.dart';
 export 'identity_overview.dart';
 export 'identity_quota.dart';
