+++ Applications/AdminUi/packages/admin_api_types/lib/src/identities/identity_deletion_process.dart
@@ -1,24 +1,9 @@
 import 'package:admin_api_types/src/admin_api_types_base.dart';
 import 'package:json_annotation/json_annotation.dart';
 
+import '../enums/enums.dart';
+
 part 'identity_deletion_process.g.dart';
-
-enum DeletionProcessStatus {
-  @JsonValue('WaitingForApproval')
-  waitingForApproval,
-
-  @JsonValue('Approved')
-  approved,
-
-  @JsonValue('Cancelled')
-  cancelled,
-
-  @JsonValue('Rejected')
-  rejected,
-
-  @JsonValue('Deleting')
-  deleting,
-}
 
 @JsonSerializable()
 class IdentityDeletionProcess {
@@ -54,22 +39,4 @@ class IdentityDeletionProcess {
 
   factory IdentityDeletionProcess.fromJson(dynamic json) => _$IdentityDeletionProcessFromJson(json as Map<String, dynamic>);
   Map<String, dynamic> toJson() => _$IdentityDeletionProcessToJson(this);
-}
-
-@JsonSerializable()
-class IdentityDeletionProcessDetail {
-  final String id;
-  final DeletionProcessStatus status;
-  final DateTime createdAt;
-  final List<IdentityDeletionProcessAuditLogEntry> auditLog;
-
-  IdentityDeletionProcessDetail({
-    required this.id,
-    required this.status,
-    required this.createdAt,
-    required this.auditLog,
-  });
-
-  factory IdentityDeletionProcessDetail.fromJson(dynamic json) => _$IdentityDeletionProcessDetailFromJson(json as Map<String, dynamic>);
-  Map<String, dynamic> toJson() => _$IdentityDeletionProcessDetailToJson(this);
 }
