+++ Applications/AdminUi/packages/admin_api_types/lib/src/messages/messages.dart
@@ -1,1 +1,2 @@
 export 'message_overview.dart';
+export 'message_recipient.dart';
